API updated. Routes are protected now, patch method is completed and users have their own lists of words which does not interfere with one another.

+++ api/controllers/dictionary.js
@@ -2,7 +2,8 @@ const mongoose = require("mongoose");
 const Words = require("../models/dictionary");
 
 exports.getAllWords = (req, res, next) => {
-  Words.find()
+  const creator = req.userData.id;
+  Words.find({ creator })
     .select("-__v")
     .exec()
     .then(data => {
@@ -29,9 +30,10 @@ exports.getAllWords = (req, res, next) => {
 };
 
 exports.getPartOfSpeech = (req, res, next) => {
+  const creator = req.userData.id;
   const partOfSpeech = req.params.partOfSpeech;
 
-  Words.find({ partOfSpeech })
+  Words.find({ creator, partOfSpeech })
     .select("-__v")
     .exec()
     .then(data => {
@@ -57,9 +59,10 @@ exports.getPartOfSpeech = (req, res, next) => {
 };
 
 exports.getOneWord = (req, res, next) => {
+  const creator = req.userData.id;
   const { partOfSpeech, name } = req.params;
 
-  Words.findOne({ partOfSpeech, name })
+  Words.findOne({ creator, partOfSpeech, name })
     .select("-__v")
     .exec()
     .then(data => {
@@ -80,15 +83,33 @@ exports.getOneWord = (req, res, next) => {
 };
 
 exports.postWord = (req, res, next) => {
-  const word = new Words({
-    _id: new mongoose.Types.ObjectId(),
-    name: req.body.name,
-    translation: req.body.translation,
-    partOfSpeech: req.body.partOfSpeech
-  });
+  const creator = req.userData.id;
+  const name = req.body.name;
 
-  word
-    .save()
+  Words.findOne({ creator, name })
+    .exec()
+    .then(word => {
+      if (word) {
+        throw new Error(
+          "Это слово уже добавлено. Вы можете обновить его перевод, если хотите."
+        );
+      }
+
+      const { name, translation, partOfSpeech } = req.body;
+
+      if (!name || !translation || !partOfSpeech) {
+        throw new Error("Для добавления слова необходимо заполнить все поля.");
+      }
+
+      const newWord = new Words({
+        name,
+        translation,
+        partOfSpeech,
+        creator: req.userData.id
+      });
+
+      return newWord.save();
+    })
     .then(data => {
       res.status(201).json({
         success: true,
@@ -104,25 +125,48 @@ exports.postWord = (req, res, next) => {
 };
 
 exports.updateWord = (req, res, next) => {
+  const creator = req.userData.id;
   const name = req.params.name;
+  const props = req.body;
 
-  res.status(200).json({
-    message: "It works!",
-    partOfSpeech,
-    name
-  });
+  function isEmpty(obj) {
+    for (const k in obj) {
+      if (obj.hasOwnProperty(k)) return false;
+    }
+    return true;
+  }
+
+  if (isEmpty(props)) {
+    throw new Error("Заполните хотя бы одно поле для изменения слова.");
+  }
+
+  Words.updateOne({ creator, name }, { $set: props })
+    .exec()
+    .then(result => {
+      res.status(200).json({
+        success: true,
+        message: `Вы успешно изменили слово ${name}!`
+      });
+    })
+    .catch(err => {
+      res.status(500).json({
+        success: false,
+        error: err
+      });
+    });
 };
 
 exports.deleteWord = (req, res, next) => {
+  const creator = req.userData.id;
   const name = req.params.name;
 
-  Words.deleteOne({ name })
+  Words.deleteOne({ creator, name })
     .select("-__v")
     .exec()
     .then(data => {
       res.status(200).json({
         success: true,
-        data
+        message: `Слово ${name} успешно удалено!`
       });
     })
     .catch(err => {
